wrap tests into a transaction (performace)
Reviewed-on: https://gerrit.tine20.org/tine20/286
Tested-by: jenkins user

+++ tests/tine20/Calendar/Controller/ResourceTest.php
@@ -33,7 +33,10 @@ class Calendar_Controller_ResourceTest extends Calendar_TestCase
     
     public function tearDown()
     {
-        Calendar_Controller_Resource::getInstance()->delete($this->_toCleanup->id);
+        if (! $this->_transactionId) {
+            Calendar_Controller_Resource::getInstance()->delete($this->_toCleanup->id);
+        }
+        
         parent::tearDown();
     }
